+++ examples/eventlog.js
@@ -1,46 +1,51 @@
 'use strict'
 
-const IpfsDaemon = require('ipfs-daemon')
+const IPFS = require('ipfs')
 const OrbitDB = require('../src/OrbitDB')
 
-const userId = Math.floor(Math.random() * 1000)
-
-const conf = { 
-  IpfsDataDir: '/tmp/' + userId,
-  Addresses: {
-    API: '/ip4/127.0.0.1/tcp/0',
-    Swarm: ['/ip4/0.0.0.0/tcp/0'],
-    Gateway: '/ip4/0.0.0.0/tcp/0'
-  },
-}
+const creatures = ['🐙', '🐷', '🐬', '🐞', '🐈', '🙉', '🐸', '🐓']
 
 console.log("Starting...")
 
-const ipfs = new IpfsDaemon(conf)
+const ipfs = new IPFS({ 
+  repo: './orbitdb/examples/ipfs',
+  start: true,
+  EXPERIMENTAL: {
+    pubsub: true,
+  },
+})
 
 ipfs.on('error', (err) => console.error(err))
 
-ipfs.on('ready', () => {
-  const orbitdb = new OrbitDB(ipfs, userId)
-  const db = orbitdb.eventlog("|orbit-db|examples|eventlog-example")
+ipfs.on('ready', async () => {
+  let db
 
-  const creatures = ['🐙', '🐷', '🐬', '🐞', '🐈', '🙉', '🐸', '🐓']
+  try {
+    const orbitdb = new OrbitDB(ipfs, './orbitdb/examples/eventlog')
+    db = await orbitdb.eventlog('example', { overwrite: true })
+  } catch (e) {
+    console.error(e)
+    process.exit(1)
+  }
 
-  const query = () => {
+  const query = async () => {
     const index = Math.floor(Math.random() * creatures.length)
-    db.add({ avatar: creatures[index], userId: userId })
-      .then(() => {
-        const latest = db.iterator({ limit: 5 }).collect()
-        let output = ``
-        output += `--------------------\n`
-        output += `Latest Visitors\n`
-        output += `--------------------\n`
-        output += latest.reverse().map((e) => e.payload.value.avatar + "  (userId: " + e.payload.value.userId + ")").join('\n') + `\n`
-        console.log(output)          
-      })
-      .catch((e) => {
-        console.error(e.stack)
-      })
+    const userId = Math.floor(Math.random() * 900 + 100)
+
+    try {
+      await db.add({ avatar: creatures[index], userId: userId })
+      const latest = db.iterator({ limit: 5 }).collect()
+      let output = ``
+      output += `[Latest Visitors]\n`
+      output += `--------------------\n`
+      output += `ID  | Visitor\n`
+      output += `--------------------\n`
+      output += latest.reverse().map((e) => e.payload.value.userId + ' | ' + e.payload.value.avatar + ')').join('\n') + `\n`
+      console.log(output)
+    } catch (e) {
+      console.error(e)
+      process.exit(1)
+    }
   }
 
   setInterval(query, 1000)