+++ examples/keyvalue.js
@@ -1,50 +1,63 @@
 'use strict'
 
-const IpfsDaemon = require('ipfs-daemon')
+const IPFS = require('ipfs')
 const OrbitDB = require('../src/OrbitDB')
 
-const userId = Math.floor(Math.random() * 1000)
+const userId = 1
+const creatures = ['🐙', '🐬', '🐋', '🐠', '🐡', '🦀', '🐢', '🐟', '🐳']
 
-const conf = { 
-  IpfsDataDir: '/tmp/' + userId,
-  Addresses: {
-    API: '/ip4/127.0.0.1/tcp/0',
-    Swarm: ['/ip4/0.0.0.0/tcp/0'],
-    Gateway: '/ip4/0.0.0.0/tcp/0'
-  },
+const output = (user) => {
+  let output = ``
+  output += `----------------------\n`
+  output += `User\n`
+  output += `----------------------\n`
+  output += `Id: ${userId}\n`
+  output += `Avatar: ${user.avatar}\n`
+  output += `Updated: ${user.updated}\n`
+  output += `----------------------\n`
+  console.log(output)
 }
 
 console.log("Starting...")
 
-const ipfs = new IpfsDaemon(conf)
+const ipfs = new IPFS({ 
+  repo: './orbitdb/examples/ipfs',
+  start: true,
+  EXPERIMENTAL: {
+    pubsub: true,
+  },
+})
 
 ipfs.on('error', (err) => console.error(err))
 
-ipfs.on('ready', () => {
-  const orbitdb = new OrbitDB(ipfs, userId)
-  const db = orbitdb.kvstore("|orbit-db|examples|kvstore-example")
-
-  const creatures = ['🐙', '🐬', '🐋', '🐠', '🐡', '🦀', '🐢', '🐟', '🐳']
-
-  const query = () => {
-    const index = Math.floor(Math.random() * creatures.length)
-    db.put(userId, { avatar: creatures[index], updated: new Date().getTime() })
-      .then(() => {
-        const user = db.get(userId)
-        let output = `\n`
-        output += `----------------------\n`
-        output += `User\n`
-        output += `----------------------\n`
-        output += `Id: ${userId}\n`
-        output += `Avatar: ${user.avatar}\n`
-        output += `Updated: ${user.updated}\n`
-        output += `----------------------`
-        console.log(output)          
-      })
-      .catch((e) => {
-        console.error(e.stack)
-      })
+ipfs.on('ready', async () => {
+  let db
+  try {
+    const orbitdb = new OrbitDB(ipfs, './orbitdb/examples/eventlog')
+    db = await orbitdb.kvstore('example', { overwrite: true })
+    await db.load()
+    // Query immediately after loading
+    const user = db.get(userId)
+    output(user)
+  } catch (e) {
+    console.error(e)
+    process.exit(1)
   }
 
+  const query = async () => {
+    // Randomly select an avatar
+    const index = Math.floor(Math.random() * creatures.length)
+
+    // Set the key to the newly selected avatar and update the timestamp
+    await db.put(userId, { avatar: creatures[index], updated: new Date().getTime() })
+
+    // Get the value of the key
+    const user = db.get(userId)
+
+    // Display the value
+    output(user)
+  }
+
+  console.log("Starting update loop...")
   setInterval(query, 1000)
 })