+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
+## v2.0.0
+- addresses
+- write-permissions
+- replication
+- performance
+- use ipfs, no more ipfs-daemon
+- async/await
+- logging
+
 ## v0.12.0
 - IPFS pubsub